# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Znaidi Mahdi Starter - Open API",
        "description": "APIs documentation usin OpenAPI",
        "contact": {
            "name": "Znaidi Mahdi",
            "email": "mahdi.znaidi@esprit.tn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "a local dev server"
        }
    ],
    "paths": {
...
        "/souscategories": {
            "get": {
                "tags": [
                    "SousCategorie"
                ],
                "summary": "List all SousCategories",
                "operationId": "indexSousCategorie",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "List All SousCategories",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "List All SousCategories",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "List All SousCategories",
                                            "type": "string",
                                            "example": "SousCategories_list"
                                        },
                                        "data": {
                                            "description": "List All SousCategories",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SousCategorie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SousCategorie"
                ],
                "summary": "Create a new SousCategorie",
                "operationId": "storeSousCategorie",
                "requestBody": {
                    "$ref": "#/components/requestBodies/SousCategorie"
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Create a new SousCategorie",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Create a new SousCategorie",
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "description": "Create a new SousCategorie",
                                            "type": "string",
                                            "example": "SousCategorie_created"
                                        },
                                        "data": {
                                            "description": "Create a new SousCategorie",
                                            "items": {
                                                "$ref": "#/components/schemas/SousCategorie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
        "/souscategories/{id}": {
            "get": {
                "tags": [
                    "SousCategorie"
                ],
                "summary": "Show an existing SousCategorie",
                "operationId": "showSousCategorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of SousCategorie to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Show an existing SousCategorie",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Show an existing SousCategorie",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "Show an existing SousCategorie",
                                            "type": "string",
                                            "example": "SousCategorie_details"
                                        },
                                        "data": {
                                            "description": "Show an existing SousCategorie",
                                            "items": {
                                                "$ref": "#/components/schemas/SousCategorie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SousCategorie"
                ],
                "summary": "Delete an existing SousCategorie",
                "operationId": "destroySousCat",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of SousCategorie to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Delete an existing SousCategorie",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Delete an existing SousCategorie",
                                            "type": "integer",
                                            "example": 204
                                        },
                                        "message": {
                                            "description": "Delete an existing SousCategorie",
                                            "type": "string",
                                            "example": "SousCategorie_deleted"
                                        },
                                        "data": {
                                            "description": "Delete an existing SousCategorie",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "SousCategorie"
                ],
                "summary": "Update an existing SousCategorie",
                "operationId": "updateSousCategories",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of equipe to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/SousCategorie"
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Update an existing SousCategorie",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Update an existing SousCategorie",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "Update an existing SousCategorie",
                                            "type": "string",
                                            "example": "SousCategorie_updated"
                                        },
                                        "data": {
                                            "description": "Update an existing SousCategorie",
                                            "items": {
                                                "$ref": "#/components/schemas/SousCategorie"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Categorie": {
                "title": "Categorie",
                "description": "Categorie model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Categorie ID",
                        "type": "integer",
                        "format": "int"
                    },
                    "libelle": {
                        "title": "string",
                        "description": "Categorie libelle",
                        "type": "string",
                        "format": "string"
                    },
                    "contrats": {
                        "title": "contrats",
                        "description": "Categorie contrat",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Contrat"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Categorie"
                }
            },
            "Contrat": {
                "title": "Contrat",
                "description": "Contrat model",
                "required": [
                    "type"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Contrat ID",
                        "type": "integer",
                        "format": "int"
                    },
                    "type": {
                        "title": "type",
                        "description": "Contrat type",
                        "type": "string",
                        "format": "string"
                    },
                    "date_debut": {
                        "title": "date_debut",
                        "description": "Experience date_debut",
                        "type": "string",
                        "format": "string"
                    },
                    "date_fin": {
                        "title": "date_fin",
                        "description": "Experience date_fin",
                        "type": "string",
                        "format": "string"
                    },
                    "actuel_salaire": {
                        "title": "salaire actuel",
                        "description": "Contrat salaire actuel",
                        "type": "number",
                        "format": "float"
                    },
                    "copie_contrat": {
                        "title": "copie_contrat",
                        "description": "Contrat copie_contrat",
                        "type": "string",
                        "format": "string"
                    },
                    "user": {
                        "title": "user",
                        "description": "Contrat user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Contrat"
                }
            },
...
            "SousCategorie": {
                "title": "SousCategorie",
                "description": "SousCategorie model",
                "required": [
                    "libelle"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "SousCategorie ID",
                        "type": "integer",
                        "format": "int"
                    },
                    "libelle": {
                        "title": "libelle",
                        "description": "SousCategorie libelle",
                        "type": "string",
                        "format": "string"
                    },
                    "categories": {
                        "title": "categories",
                        "description": "Categorie",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Categorie"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SousCategorie"
                }
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "User id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "type": "string",
                        "format": "string"
                    },
                    "matricule_hr": {
                        "title": "matricule_hr",
                        "description": "User matricule_hr",
                        "type": "string",
                        "format": "string"
                    },
                    "nom": {
                        "title": "nom",
                        "description": "User nom",
                        "type": "string",
                        "format": "string"
                    },
                    "prenom": {
                        "title": "prenom",
                        "description": "User prenom",
                        "type": "string",
                        "format": "string"
                    },
                    "adresse": {
                        "title": "adresse",
                        "description": "User adresse",
                        "type": "string",
                        "format": "string"
                    },
                    "num_telephone": {
                        "title": "numéro_téléphone",
                        "description": "User numéro_telephone",
                        "type": "string",
                        "format": "string"
                    },
                    "cin_passport": {
                        "title": "cin_passport",
                        "description": "User cin_passport",
                        "type": "string",
                        "format": "string"
                    },
                    "etat_civil": {
                        "title": "etat_civil",
                        "description": "User etat_civil",
                        "type": "string",
                        "format": "string"
                    },
                    "image": {
                        "title": "image",
                        "description": "User image",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "SousCategorieRequestBody": {
                "title": "SousCategorieRequestBody",
                "description": "SousCategorie model",
                "properties": {
                    "libelle": {
                        "title": "string",
                        "description": "SousCategorie libelle",
                        "type": "string",
                        "format": "string"
                    },
                    "categories": {
                        "title": "string",
                        "description": "categories id",
                        "type": "integer",
                        "format": "int"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Contrat"
                }
            },
...
        },
        "requestBodies": {
...
            "SousCategorie": {
                "description": "SousCategorie object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/SousCategorieRequestBody"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "Password Based": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}